Fixing clippy on Mac.
