+++ src/macos/common.rs
@@ -1,0 +1,72 @@
+use cocoa::base::id;
+use core_graphics::event::{CGEvent, CGEventTapLocation, CGEventType};
+use std::os::raw::c_void;
+
+pub type CFMachPortRef = *const c_void;
+pub type CFIndex = u64;
+pub type CFAllocatorRef = id;
+pub type CFRunLoopSourceRef = id;
+pub type CFRunLoopRef = id;
+pub type CFRunLoopMode = id;
+pub type CGEventTapProxy = id;
+pub type CGEventRef = CGEvent;
+
+// https://developer.apple.com/documentation/coregraphics/cgeventtapplacement?language=objc
+pub type CGEventTapPlacement = u32;
+#[allow(non_upper_case_globals)]
+pub const kCGHeadInsertEventTap: u32 = 0;
+
+// https://developer.apple.com/documentation/coregraphics/cgeventtapoptions?language=objc
+#[allow(non_upper_case_globals)]
+#[repr(u32)]
+pub enum CGEventTapOption {
+    Default = 0,
+    ListenOnly = 1,
+}
+
+// https://developer.apple.com/documentation/coregraphics/cgeventmask?language=objc
+pub type CGEventMask = u64;
+#[allow(non_upper_case_globals)]
+pub const kCGEventMaskForAllEvents: u64 = (1 << CGEventType::LeftMouseDown as u64)
+    + (1 << CGEventType::LeftMouseUp as u64)
+    + (1 << CGEventType::RightMouseDown as u64)
+    + (1 << CGEventType::RightMouseUp as u64)
+    + (1 << CGEventType::MouseMoved as u64)
+    + (1 << CGEventType::LeftMouseDragged as u64)
+    + (1 << CGEventType::RightMouseDragged as u64)
+    + (1 << CGEventType::KeyDown as u64)
+    + (1 << CGEventType::KeyUp as u64)
+    + (1 << CGEventType::FlagsChanged as u64)
+    + (1 << CGEventType::ScrollWheel as u64);
+
+#[cfg(target_os = "macos")]
+#[link(name = "Cocoa", kind = "framework")]
+extern "C" {
+    #[allow(improper_ctypes)]
+    pub fn CGEventTapCreate(
+        tap: CGEventTapLocation,
+        place: CGEventTapPlacement,
+        options: CGEventTapOption,
+        eventsOfInterest: CGEventMask,
+        callback: QCallback,
+        user_info: id,
+    ) -> CFMachPortRef;
+    pub fn CFMachPortCreateRunLoopSource(
+        allocator: CFAllocatorRef,
+        tap: CFMachPortRef,
+        order: CFIndex,
+    ) -> CFRunLoopSourceRef;
+    pub fn CFRunLoopAddSource(rl: CFRunLoopRef, source: CFRunLoopSourceRef, mode: CFRunLoopMode);
+    pub fn CFRunLoopGetCurrent() -> CFRunLoopRef;
+    pub fn CGEventTapEnable(tap: CFMachPortRef, enable: bool);
+    pub fn CFRunLoopRun();
+
+    pub static kCFRunLoopCommonModes: CFRunLoopMode;
+
+}
+pub type QCallback = unsafe extern "C" fn(
+    proxy: CGEventTapProxy,
+    _type: CGEventType,
+    cg_event: CGEventRef,
+    user_info: *mut c_void,
+) -> CGEventRef;
